separate actions->code logic into new source
